refactor: Rename to RudderDataResidencyManager
